A prototype implementation of etcd-based fault tolerance for the master (#7)
* Implement fault-tolerant master with etcd

This implements a master module that keeps all its state in etcd, so
that it can be killed and restarted without disruption to the chain.

It responds to the same commands as the current master, except that
`MASTER.REMOVE` is fully implemented, including for middle nodes.

It detects failures by watching for heartbeat keys in etcd, which are
published with a chain-specific prefix and are just the connection info
of the relevant node. This is implemented in `heartbeat_monitor.cc`.

Members publish this key when processing the `MEMBER.CONNECT_TO_MASTER`
command, and an aeTimeEvent on their event loop refreshes the TTL of the
key on a configurable interval.

However, the client can notify the master that a node died by calling
`MASTER.REMOVE` with the relevant connection info too.

The member has been updated to support the etcd master with an
alternative master mode (`MasterMode::kEtcd`).

On connecting, it writes its connection info to a key as described
above.  If the process fails to send a KeepAlive request to etcd within
the time limit, the key disappears and the master considers the node to
have crashed.

The head and tail read and write checkpoints directly into etcd, rather
than the master itself, so that this data is also replicated.

To avoid name collision issues between the stored state of various
master services, the master connection info is now a URL
(host:port/path), with the Redis-style master simply having no path:

```
MEMBER.CONNECT_TO_MASTER 127.0.0.1:6369 # Connecting to a Redis master
MEMBER.CONNECT_TO_MASTER 127.0.0.1:2379/chain_name # Connecting to an etcd master
```

* Remove nlohmann single-header to add as a submodule

* Add nlohmann/json as a git submodule

* Remove duplicate argument parsing logic

* PR comments:

1. clang-format on all modified .cc/.h files
2. Documentation touchups
3. Move utils::* to global namespace
4. Represent EtcdURL as a struct, not a class

* Standardize indentation in CMakeLists

* Add instructions to install Protobuf

* Rename nlohmann submodule to nlohmann-json

* Use same name as repo name for etcd3-cpp

* Manually remove blank line that clang-format wouldn't

* Re-add missing build instructions

* Some fixes to build instructions

* Add -fPIC to etcd lib compile flags

* Update README.md

* Link etcd3 lib with protobufs

* Rename mode getters/setters to fix conflict

* Small fixes to make tests run and pass on Ubuntu

* Remove unused unix socket def. From old impl...

* Rearrange args to work with BSD pkill

* Check each chain member when master recovers

* Build changes needed

* FindLevelDB fix

* Fix style issues in etcd utils

* Style fixes

* Add readthis